Relocate the logging middleware
